+++ app/forms/CreateMenuForm.scala
@@ -3,4 +3,4 @@ package forms
 /**
   * Created by Hasi on 21.11.2016.
   */
-case class CreateMenuForm(name: String, price: Int, category: String)
+case class CreateMenuForm(name: String, price: Double, category: String)
