+++ app/models/Menu.scala
@@ -3,4 +3,4 @@ package models
 /**
   * Created by Hasi on 21.11.2016.
   */
-case class Menu(var id: Long, var name: String, var price: Int, var category: String)
+case class Menu(var id: Long, var name: String, var price: Double, var category: String)
